Added hook useModalFocus
