+++ src/ui/Modal/Modal.js
@@ -1,32 +1,16 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
 
 import { Dialog } from '@ufx-ui/core'
+import useModalFocus from '../../hooks/useModalFocus'
 import Scrollbars from '../Scrollbars'
 
 import './style.css'
 
-const FOCUSABLE_ELEMENTS = ['input', 'button', '[role=button]']
-
 const Modal = ({
   label, isOpen, onClose, onSubmit, children, className, scrollable, ...rest
 }) => {
-  useEffect(() => {
-    // focus on the first interactable element
-    if (isOpen) {
-      // eslint-disable-next-line lodash/prefer-lodash-method
-      const el = document.querySelector(FOCUSABLE_ELEMENTS.map(element => `.modal__body ${element}`).join(','))
-      const footer = document.querySelector('.modal__footer')
-      if (el && footer && !footer.contains(el)) {
-        el.focus()
-      } else {
-        const titleNode = document.querySelector('.modal__title')
-        if (titleNode && titleNode.focus) {
-          titleNode.focus()
-        }
-      }
-    }
-  }, [isOpen])
+  useModalFocus()
 
   return (
     <Dialog
